backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
joblib==1.4.2
numpy==2.2.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.6.0
urllib3==2.4.0
